contract: Extract contract schemas
