+++ apps/backend/package.json
@@ -16,10 +16,11 @@
     "@repo/contract": "workspace:*",
     "zod": "^3.23.8",
     "fastify": "^5.0.0",
-    "@fastify/static": "^8.0.1",
+    "@scalar/fastify-api-reference": "^1.25.17",
     "@ts-rest/core": "^3.51.0",
     "@ts-rest/fastify": "^3.51.0",
     "@ts-rest/open-api": "^3.51.0",
+
     "@paralleldrive/cuid2": "^2.2.2",
     "normalize-url": "^8.0.1",
     "pino": "^9.3.2",
