+++ apps/backend/src/index.ts
@@ -1,10 +1,3 @@
-// import { routes } from './routes/plugin';
-// import { logger } from './logger';
-// import { logger as honoLogger } from 'hono-pino';
-// import { cors } from 'hono/cors';
-// import { OpenAPIHono } from '@hono/zod-openapi';
-// import { apiReference } from '@scalar/hono-api-reference';
-
 import { contract } from '@repo/contract';
 import { initServer } from '@ts-rest/fastify';
 import { fastify } from 'fastify';
@@ -12,30 +5,8 @@ import { authRouter } from './routes/auth';
 import { geojsonRouter } from './routes/geojson';
 import { spritesRouter } from './routes/sprites';
 import { stylesRouter } from './routes/styles';
-
-// export const app = new OpenAPIHono()
-//   .doc31('/openapi.json', {
-//     openapi: '3.1.0',
-//     info: {
-//       version: '1.0.0',
-//       title: 'My API',
-//     },
-//   })
-//   .get('/docs', apiReference({ spec: { url: '/openapi.json' }, theme: 'kepler' }))
-//   .use(
-//     honoLogger({
-//       pino: logger,
-//     }),
-//   )
-//   .use(cors())
-//   .route('', routes);
-
-// export type AppType = typeof app;
-
-// export default {
-//   port: 8000,
-//   fetch: app.fetch,
-// };
+import { generateOpenApi } from '@ts-rest/open-api';
+import ScalarApiReference from '@scalar/fastify-api-reference';
 
 const s = initServer();
 const router = s.router(contract, {
@@ -46,5 +17,27 @@ const router = s.router(contract, {
 });
 
 const app = fastify();
+// Register ts-rest routes
 s.registerRouter(contract, router, app);
+// OpenAPI schema
+app.get('/openapi.json', async (req, reply) => {
+  return reply.send(
+    generateOpenApi(contract, {
+      info: {
+        title: 'TrailEyes API',
+        version: '1.0.0',
+      },
+    }),
+  );
+});
+app.register(ScalarApiReference, {
+  routePrefix: '/docs',
+  configuration: {
+    spec: {
+      url: '/openapi.json',
+    },
+    theme: 'kepler',
+  },
+});
+
 await app.listen({ port: 8000 });
