+++ packages/contract/src/models/auth.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod';
+
+export const LoginResponseSchema = z.object({
+  requiresSecondFactor: z.boolean(),
+  enabledSecondFactors: z.array(z.string()),
+});
