+++ packages/contract/src/models/base.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+
+export const ErrorResponseBaseSchema = z.object({
+  statusCode: z.number(),
+  message: z.string(),
+  error: z.string(),
+  code: z.string().optional(),
+});
